Convert BaseAvatar and UserAvatar to TypeScript
